+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## New version
 - Add project name validator
+- Add commander
 
 ## v.2022.11.11
 - Add CODE_OF_CONDUCT file