+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # tsci changes
 
+ ## 2023.1.2
+- Fix dependabot issues
  ## 2022.12.05
 - Add project name validator
 - Add commander